% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grt_base.R
\name{anova.grt}
\alias{anova.grt}
\title{Compare nested GRT models}
\usage{
\method{anova}{grt}(model1, model2)
}
\arguments{
\item{model1}{A fitted GRT model returned by fit.grt}

\item{model2}{A larger GRT model, with model1 nested inside}
}
\description{
Conducts a likelihood-ratio G-test on nested GRT models.
}

