\name{plot.grt}
\alias{plot.grt}
\title{plot.grt}
\usage{
  \method{plot}{grt} (x, xlab = NULL, ylab = NULL, ...)
}
\arguments{
  \item{x}{a grt object, as returned by fit.grt}

  \item{xlab}{optional label for the x axis (defaults to
  NULL)}

  \item{ylab}{optional label for the y axis (defaults to
  NULL)}
}
\description{
  Plot the object returned by fit.grt
}

