\name{print.grt}
\alias{print.grt}
\title{print.grt}
\usage{
  \method{print}{grt} (x, ...)
}
\arguments{
  \item{x}{An object returned by fit.grt}

  \item{...}{further arguments passed to or from other
  methods, as in the generic print function}
}
\description{
  Print the object returned by fit.grt
}

