\name{silbert09a}
\alias{silbert09a}
\title{Frequency vs. Duration confusion matrix}
\format{A \code{matrix} instance, containing counts for all stimulus-response combinations. Rows correspond to stimuli, columns to responses}
\source{
  Silbert, N. H., Townsend, J. T., & Lentz, J. J. (2009).
  Independence and separability in the perception of
  complex nonspeech sounds. Attention, Perception, &
  Psychophysics, 71(8), 1900-1915.
}
\description{
  Confusion matrix from auditory perception experiment, in
  which listeners identified noise stimuli varying across
  frequency range and duration (Experiment 1, Observer 3 in
  Ref.)
}
\author{
  Noah H. Silbert
}

