\name{silbert09b}
\alias{silbert09b}
\title{Pitch vs. Timbre confusion matrix}
\format{A \code{matrix} instance, containing counts for all stimulus-response combinations. Rows correspond to stimuli, columns to responses}
\source{
  Silbert, N. H., Townsend, J. T., & Lentz, J. J. (2009).
  Independence and separability in the perception of
  complex nonspeech sounds. Attention, Perception, &
  Psychophysics, 71(8), 1900-1915.
}
\description{
  Confusion matrix from auditory perception experiment, in
  which listeners identified 13-component harmonic stimuli
  varying across fundamental frequency and location of
  spectral prominence (Experiment 2, Observer 7 in Ref..
}
\author{
  Noah H. Silbert
}

