% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grt-data.R
\name{silbert12}
\alias{silbert12}
\title{Phoneme confusion matrix}
\format{A \code{matrix} instance, containing counts for all stimulus-response combinations. Rows correspond to stimuli, columns to responses}
\source{
Silbert, N. H. (2012). Syllable structure and integration of voicing and manner of articulation information in labial consonant identification. Journal of the Acoustical Society of America, 131(5), 4076-4086.
}
\usage{
data(silbert12)
}
\description{
Confusion matrix from speech perception experiment probing confusions between noise-masked tokens of English [p],[b],[f], and [v] (observer 3 in Ref.)
}
\author{
Noah H. Silbert
}

