\name{summary.grt}
\alias{summary.grt}
\title{summary.grt}
\usage{
  \method{summary}{grt} (object, ...)
}
\arguments{
  \item{object}{An object retrned by fit.grt}

  \item{...}{additional arguments affecting the summary
  produced, as in the generic summary function}
}
\description{
  Summarize the object returned by fit.grt
}

