% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grt-data.R
\docType{data}
\name{thomas15b}
\alias{thomas15b}
\title{3x3 face recognition confusion matrix for Observer B}
\format{an \code{xtabs} instance, containing counts for all stimulus-response combinations.
The first two dimensions consist of the response counts on each level of nose width and eye separation, respectively, and the third dimension indexes the stimulus.}
\source{
Thomas, R. D., Altieri, N. A., Silbert, N. H., Wenger, M. J., & Wessels, P. M. (2015). Multidimensional signal detection decision models of the uncertainty task: Application to face perception. Journal of Mathematical Psychology, 66, 16-33.
}
\usage{
data(thomas15b)
}
\description{
This data set contains the results of a 3x3 full-report face recognition experiment reported in Thomas et al (2015). The two channels are degree of eye separation and nose width, with three levels on each dimension.
}
\author{
Robin D. Thomas
}

