\docType{data}
\name{thomasB}
\alias{thomasB}
\title{Face recognition confusion matrix for Observer B}
\format{a \code{matrix} instance, containing counts for all stimulus-response combinations. Each row corresponds to a different stimulus presentation (in the order aa, ab, ba, bb) and each column in that row represents the frequency of each response (in the order aa, ab, ba, bb).}
\source{
  Thomas, R. D. (2001). Characterizing perceptual
  interactions in face identification using
  multidimensional signal detection theory. In M.Wenger &
  J.T. Townsend (Eds.) Computational, geometric, and
  process perspectives on facial cognition: Contexts and
  challenges. Hillsdale, NJ: Erlbaum.
}
\description{
  This data set contains the results of a full-report face
  recognition experiment reported in Thomas (2001). For
  Observer B, the two channels are degree of eye separation
  and mouth width.
}
\author{
  Robin D. Thomas
}
\keyword{datasets}

