\name{plot.grt}
\alias{plot.grt}
\title{Plot the object returned by fit.grt}
\usage{
  \method{plot}{grt} (x, level = 0.5, xlab = NULL,
    ylab = NULL, ...)
}
\arguments{
  \item{x}{a grt object, as returned by fit.grt}

  \item{level}{number between 0 and 1 indicating which
  contour to plot (defaults to .5)}

  \item{xlab}{optional label for the x axis (defaults to
  NULL)}

  \item{ylab}{optional label for the y axis (defaults to
  NULL)}

  \item{...}{Arguments to be passed to methods, as in
  generic plot function}
}
\description{
  Plot the object returned by fit.grt
}

