% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grt_base.R
\name{summary.grt}
\alias{summary.grt}
\title{summary.grt}
\usage{
\method{summary}{grt}(object, ...)
}
\arguments{
\item{object}{An object returned by fit.grt}

\item{...}{additional arguments affecting the summary produced, as in the generic summary function}
}
\description{
Summarize the object returned by fit.grt
}

